refactor:refactor FavouriteScreen.kt
