+++ composeApp/src/commonMain/kotlin/com/cairosquad/evolvefit/ui/screen/favorites/FavoritesScreen.kt
@@ -19,8 +19,9 @@ import com.cairosquad.evolvefit.design_system.component.appbar.ActionIconButton
 import com.cairosquad.evolvefit.design_system.component.appbar.CustomAppBar
 import com.cairosquad.evolvefit.design_system.theme.AppTheme
 import com.cairosquad.evolvefit.design_system.theme.Theme
-import com.cairosquad.evolvefit.ui.screen.favorites.componants.TabsWithPager
-import com.cairosquad.evolvefit.ui.screen.favorites.componants.dummyWorkouts
+import com.cairosquad.evolvefit.ui.screen.favorites.content.componants.TabsWithPager
+import com.cairosquad.evolvefit.ui.screen.favorites.content.componants.dummyWorkouts
+import com.cairosquad.evolvefit.ui.screen.favorites.content.FavoritesLoadingScreen
 import com.cairosquad.evolvefit.ui.util.ObserveAsEffect
 import com.cairosquad.evolvefit.viewmodel.favorites.FavoritesEffect
 import com.cairosquad.evolvefit.viewmodel.favorites.FavoritesInteractionListener
